+++ todo_tig169/lib/main.dart
@@ -1,11 +1,15 @@
+import 'package:exercise_ui_02/model.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 
-import './SecondView.dart';
-import './model.dart';
-import './Constants.dart';
+import './MainView.dart';
 
 void main() {
-  runApp(MyApp());
+  var state = MyState();
+  runApp(ChangeNotifierProvider(
+    create: (context) => state,
+    child: MyApp(),
+  ));
 }
 
 class MyApp extends StatelessWidget {
@@ -15,93 +19,4 @@ class MyApp extends StatelessWidget {
       home: MainView(),
     );
   }
-}
-
-class TodoList extends StatelessWidget {
-  final List<TodoItem> list;
-
-  TodoList(this.list);
-
-  Widget build(BuildContext context) {
-    return ListView(children: list.map((item) => _item(item)).toList());
-  }
-
-  Widget _item(item) {
-    return ListTile(
-      leading: Checkbox(
-        activeColor: Colors.deepPurple,
-        value: false,
-        onChanged: (val) {},
-      ),
-      title: Text(
-        item.text,
-        style: TextStyle(fontSize: 20, color: Colors.deepPurple),
-      ),
-      trailing: IconButton(
-          icon: Icon(Icons.clear), color: Colors.deepPurple, onPressed: () {}),
-    );
-  }
-}
-
-class MainView extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-          flexibleSpace: _appBarDecoration(),
-          title: Text("TIG169 TODO"),
-          centerTitle: true,
-          actions: [
-            PopupMenuButton<String>(
-                onSelected: choiceAction,
-                itemBuilder: (BuildContext context) {
-                  return Constants.choices.map((String choice) {
-                    return PopupMenuItem<String>(
-                      value: choice,
-                      child: Text(choice),
-                    );
-                  }).toList();
-                })
-          ],
-        ),
-        body: TodoList([
-          TodoItem(text: "Drink tea"),
-          TodoItem(text: "Drink more tea"),
-          TodoItem(text: "Drink a pot of tea")
-        ]),
-        floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            Navigator.push(
-                context,
-                MaterialPageRoute(
-                    builder: (context) => SecondView(TodoItem(text: ""))));
-          },
-          tooltip: 'Increment',
-          backgroundColor: Colors.deepPurple[300],
-          child: Icon(Icons.add),
-        ));
-  }
-
-  Widget _appBarDecoration() {
-    return Container(
-      decoration: BoxDecoration(
-        image: DecorationImage(
-          image: AssetImage("assets/purple_sky.jpg"),
-          fit: BoxFit.cover,
-        ),
-      ),
-    );
-  } //_appBarDecoration, decorates the appbar with an image
-
-  void choiceAction(String choice) {
-    if (choice == Constants.All) {
-      print("All");
-    }
-    if (choice == Constants.Done) {
-      print("Done");
-    }
-    if (choice == Constants.UnDone) {
-      print("Undone");
-    }
-  } //prints in the debug when pressed.
 }
