+++ todo_tig169/lib/model.dart
@@ -5,3 +5,14 @@ class TodoItem {
 
   TodoItem({this.text});
 } //TodoItem, creates a TodoItem.
+
+class MyState extends ChangeNotifier {
+  List<TodoItem> _list;
+
+  List<TodoItem> get list => _list;
+
+  void addItem(TodoItem item) {
+    _list.add(item);
+    notifyListeners();
+  }
+}
