+++ todo_tig169/pubspec.yaml
@@ -27,7 +27,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-
+  provider: ^4.3.2+2
 dev_dependencies:
   flutter_test:
     sdk: flutter
